% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmpiricalCalibration.R
\name{doEmpiricalCalibration}
\alias{doEmpiricalCalibration}
\title{Perform empirical calibration}
\usage{
doEmpiricalCalibration(workFolder, study)
}
\arguments{
\item{workFolder}{The path to the output folder containing the results.}
}
\description{
Perform empirical calibration
}
\details{
Performs empirical calibration of confidence intervals and p-values using the negative and positive
control outcomes.
}
