% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageResultsForSharing.R
\name{packageResults}
\alias{packageResults}
\title{Package the results for sharing with OHDSI researchers}
\usage{
packageResults(outputFolder)
}
\arguments{
\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes (/)}
}
\description{
Package the results for sharing with OHDSI researchers
}
\details{
This function packages the results.
}
