% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortMethod.R
\name{runCohortMethodGraham}
\alias{runCohortMethodGraham}
\title{Run the Graham study replication}
\usage{
runCohortMethodGraham(connectionDetails, cdmDatabaseSchema,
  workDatabaseSchema = cdmDatabaseSchema,
  studyCohortTable = "ohdsi_ci_calibration", oracleTempSchema = NULL,
  maxCores = 4)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{workDatabaseSchema}{Schema name where intermediate data can be stored. You will need to have
write priviliges in this schema. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{studyCohortTable}{The name of the table that will be created in the work database schema.
This table will hold the exposure and outcome cohorts used in this
study.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
priviliges for storing temporary tables.}

\item{maxCores}{How many parallel cores should be used? If more cores are made available
this can speed up the analyses.}
}
\description{
Run the Graham study replication
}
